♻️ Refactor: 코드래빗 리뷰 반영 (페이징 정렬 안정화, users LEFT JOIN, offset 오버플로우 방지)

+++ proovy/src/main/java/com/deundeun/challenge/dto/request/ChallengeSearchCondition.java
@@ -15,8 +15,8 @@ public record ChallengeSearchCondition(
         if (size == null || size < 1 || size > 50)size = 10;
     }
 
-    public int offset() {
-        return page * size;
+    public long offset() {
+            return (long) page * size;
     }
 
 }
